requirements.txt
black==24.10.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
graphviz==0.20.3
importlib_resources==6.4.5
itermplot==0.5
kiwisolver==1.4.7
matplotlib==3.9.2
mypy-extensions==1.0.0
numpy==2.0.2
packaging==24.1
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pyparsing==3.2.0
python-dateutil==2.9.0.post0
six==1.16.0
tomli==2.0.2
typing_extensions==4.12.2
zipp==3.20.2
